USB: Test and fix errors on x86 boxes
Added support for more than one UHCI controller. Make sure the 'PIRQ' bit in the
legacy support register (PCI config space) of the UHCI controller is enabled and
also that the 'Trap on IRQ' bit is disabled. Fix offset bug in PCI-config space
emulation for Linux.

Fixes #282

+++ dde_linux/src/drivers/usb/arm/platform/platform.cc
@@ -398,7 +398,7 @@ void platform_hcd_init(Services *services)
 	/*
 	 * Needed for DMA buffer allocation. See 'hcd_buffer_alloc' in 'buffer.c'
 	 */
-	static u64 dma_mask = ~(u32)0;
+	static u64 dma_mask = ~(u64)0;
 	pdev->dev.dma_mask = &dma_mask;
 	pdev->dev.coherent_dma_mask = ~0;
